requirements.txt
flask==2.2.2
python-dotenv==0.21.0
spotipy==2.20.0
boto3==1.24.96